requirements.txt
dash==1.13.3
dash-core-components==1.10.1
dash-html-components==1.0.3
pandas==1.0.5
numpy==1.18.1
plotly==4.14.3
gunicorn==20.1.0
flask==1.1.2
flask-restful==0.3.9
imdbpy==2021.4.18
wikipedia==1.4.0
